+++ src/node/index.ts
@@ -1,4 +1,5 @@
 export * from './server'
 export * from './build'
+export * from './depOptimizer'
 export * from './config'
 export { readBody, cachedRead, isStaticAsset, isImportRequest } from './utils'